fixed a little issue where deleting local storage manually wouldn't affect the app except by reloading the page, now using useLocation from react router to check if token exist at each change of route
